Restructure code to core and node directories
This builds a platform for the web version of the sdk.
This abstracts out the grpc related code to nodejs and keeps the base sdk and service client grpc agnostic.
