+++ src/core/PaymentChannel.js
@@ -1,7 +1,6 @@
 import BigNumber from 'bignumber.js';
 
-import paymentChannelStateMessages from './payment_channel_state_service_pb';
-import logger from './utils/logger';
+import logger from '../utils/logger';
 
 class PaymentChannel {
   /**
@@ -16,7 +15,7 @@ class PaymentChannel {
     this._web3 = web3;
     this._account = account;
     this._mpeContract = mpeContract;
-    this._paymentChannelStateServiceClient = service.paymentChannelStateServiceClient;
+    this._serviceClient = service;
     this._state = {
       nonce: new BigNumber(0),
       lastSignedAmount: new BigNumber(0),
@@ -88,33 +87,21 @@ class PaymentChannel {
   }
 
   async _currentChannelState() {
-    const signatureBytes = this._account.signedData({ t: 'uint256', v: this.channelId });
-
-    const channelIdBytes = Buffer.alloc(4);
-    channelIdBytes.writeUInt32BE(this.channelId, 0);
-
-    const channelStateRequest = new paymentChannelStateMessages.ChannelStateRequest();
-    channelStateRequest.setChannelId(channelIdBytes);
-    channelStateRequest.setSignature(signatureBytes);
-
     logger.debug(`Fetching latest PaymentChannel[id: ${this.channelId}] state from service daemon`, { tags: ['PaymentChannel'] });
-    return new Promise((resolve, reject) => {
-      this._paymentChannelStateServiceClient.getChannelState(channelStateRequest, (err, response) => {
-        if(err) {
-          logger.error(`Failed to fetch latest PaymentChannel[id: ${this.channelId}] state from service daemon. ${err.message}`, { tags: ['PaymentChannel'] });
-          reject(err);
-        } else {
-          const nonce = PaymentChannel._uint8ArrayToBN(response.getCurrentNonce());
-          const currentSignedAmount = PaymentChannel._uint8ArrayToBN(response.getCurrentSignedAmount());
-          const channelState = {
-            lastSignedAmount: currentSignedAmount,
-            nonce,
-          };
-          logger.debug(`Latest PaymentChannel[id: ${this.channelId}] state: {lastSignedAmount: ${currentSignedAmount}, nonce: ${nonce}}`, { tags: ['PaymentChannel'] });
-          resolve(channelState);
-        }
-      });
-    });
+    try {
+      const response = await this._serviceClient.getChannelState(this.channelId);
+      const nonce = PaymentChannel._uint8ArrayToBN(response.getCurrentNonce());
+      const currentSignedAmount = PaymentChannel._uint8ArrayToBN(response.getCurrentSignedAmount());
+      const channelState = {
+        lastSignedAmount: currentSignedAmount,
+        nonce,
+      };
+      logger.debug(`Latest PaymentChannel[id: ${this.channelId}] state: {lastSignedAmount: ${currentSignedAmount}, nonce: ${nonce}}`, { tags: ['PaymentChannel'] });
+      return Promise.resolve(channelState);
+    } catch(err) {
+      logger.error(`Failed to fetch latest PaymentChannel[id: ${this.channelId}] state from service daemon. ${err.message}`, { tags: ['PaymentChannel'] });
+      return Promise.reject(err);
+    }
   }
 
   static _uint8ArrayToBN(uint8Array) {
