+++ src/core/payment_channel_management_strategies/DefaultPaymentChannelManagementStrategy.js
