+++ src/nodejs/index.js
@@ -1,0 +1,6 @@
+import DefaultPaymentChannelManagementStrategy from '../core/payment_channel_management_strategies';
+import NodeSdk from './NodeSdk';
+
+export { DefaultPaymentChannelManagementStrategy };
+
+export default NodeSdk;
