+++ src/nodejs/proto/payment_channel_state_service_pb.js
